Level 1: Implement add and list functions
Implemented addTask(String), which adds a task inputted by the user to
an ArrayList. This is now the default behaviour of the bot, if the
command inputted does not match any other commands.
Implemented printTasks(), which prints all stored tasks to stdout.
echo() still exists, but is obsolete and unused.