Refactor Duke.java to be split into multiple methods
Huge update. Introduction of enums for faster comparison than Strings
and to better take advantage of switch statements.
This also makes it extremely easy to add new commands

Add HELP command also for better user experience

Add many functions to handle edge cases.
The bot now tells the user when the command is invalid and sometimes why.
The bot does not break when commands are inputted not expected
Overall improvement of user experience.